Unusual Method
I've began manipulating the data from our string list.
Currently it tests each element as to what type it is (application,
process, io, or system) and manipulates it accordingly.
The application tests if it is starting or ending. If its ending that
means we can now store the entire application into our appList. Not
completed yet, but for processes a similar thing should happen in that
until another process or application starts, a process should continue
taking in information (IO). Its confusing how I'm doing it right now, if
you have questions on it let me know.