Load configuration file
Wasn't sure if you guys had done it yet, but I did the code to load the
configuration file to the global variables. The configuration file
should be an argument that gets passed through the command line and in
order to simulate that through visual studios you'll have to go through
these steps and add it as a command argument (config_1.txt):

To access the Debug page, select a project node in Solution Explorer. On
the Project menu, choose ProjectName Properties. When the Project
Designer appears, click the Debug tab.

Also the configuration file should be where your project is stored. This
is where mine is for it to load:

C:\Users\Jennifer\Documents\Visual Studio
2013\Projects\ConsoleApplication1\ConsoleApplication1\bin\Debug

It'll probably change later when we submit our project. Not sure how it
works yet when we run it by command line.