Update currentProg.txt
This is my method of loading all our data into the application list. In the end it prints the structure of our fully loaded application list including its processes and their corresponding IO. 
NOTE: WE SHOULD BE USING CONSOLE APPLICATION.
I noticed in program.cs (11/23/14 at1:11 am) we are using namespace simulator. I don't know if that's a huge difference, but I suggest we use a console application to do this project.