Combined changes to program
Namespaces are C# program elements designed to help you organize your
programs. They also provide assistance in avoiding name clashes between
two sets of code. Implementing Namespaces in your own code is a good
habit because it is likely to save you from problems later when you want
to reuse some of your code. For example, if you created a class named
Console, you would need to put it in your own namespace to ensure that
there wasn't any confusion about when the System.Console class should be
used or when your class should be used. Generally, it would be a bad
idea to create a class named Console, but in many cases your classes
will be named the same as classes in either the .NET Framework Class
Library or a third party library and namespaces help you avoid the
problems that identical class names would cause.