Adding if statements to sort data based on scheduler
The only new additions are 2 if statements for if the scheduler is SJF
or SRJN and the main loop. This loop will run while the bool
allAppsFinished is not true. We will probably need to write a function
to check if all apps are finished periodically and if they are the
function will set that bool to true.